requirements.txt
asgiref==3.7.2
awscli>=1.32.69,<2.0
boto3>=1.34.95,<2.0
Django==4.2.11
django-cities-light==3.10.1
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==24.2
django-import-export==4.1.1
django-phonenumber-field==7.3.0
django-redis==5.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==4.7.2
django-storages==1.14.2
drf-yasg==1.21.7
exceptiongroup==1.2.0
greenlet==3.0.3
iniconfig==2.0.0
mixpanel==4.10.1
packaging==23.2
phonenumbers==8.13.31
pillow==10.3.0
pluggy==1.4.0
psycopg2-binary==2.9.9
pytest==8.0.2
pytest-django==4.8.0
pytest-mock==3.14.0
python-dotenv==1.0.1
pytz==2024.1
sendgrid==6.9.7
setuptools>=65.0.0
stripe==4.2.0
sqlparse==0.4.4
tomli==2.0.1
twilio==9.1.0
typing_extensions==4.10.0
flake8==7.0.0
